FreeImage elimiado y materiales arreglados sin basura

+++ P3/ProyectoBase/Src/Engine/TestComponent.cpp
@@ -50,7 +50,7 @@ public:
 		TestComponent* testComponent = new TestComponent();
 		testComponent->setFather(father);
 		testComponent->setMaterial(_data["material"].asString());
-		testComponent->setColor(Ogre::Vector3(_data["color"][0].asInt(), _data["color"][1].asInt(), _data["color"][2].asInt()));
+		testComponent->setColor(Ogre::Vector3(_data["color"][0].asFloat(), _data["color"][1].asFloat(), _data["color"][2].asFloat()));
 
 		componentManager->addPC(testComponent);
 		return testComponent;
